move multiplexer inside store4
