+++ store/src/main/java/com/nytimes/android/external/store4/FetcherController.kt
@@ -1,6 +1,6 @@
 package com.nytimes.android.external.store4
 
-import com.nytimes.android.external.store3.multiplex.Multiplexer
+import com.nytimes.android.external.store4.multiplex.Multiplexer
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CoroutineStart
@@ -30,6 +30,8 @@ internal class FetcherController<Key, Input, Output>(
     fun getFetcher(key: Key): Flow<Input> {
         val multiplexer = multiplexerLock.withLock {
             fetchers.getOrPut(key) {
+                // TODO
+                //  we need cleanup here
                 Multiplexer(
                     scope = scope,
                     bufferSize = 0,
