+++ store/src/main/java/com/nytimes/android/external/store4/multiplex/Multiplexer.kt
@@ -1,4 +1,4 @@
-package com.nytimes.android.external.store3.multiplex
+package com.nytimes.android.external.store4.multiplex
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
