+++ store/src/test/java/com/nytimes/android/external/store4/multiplex/MultiplexTest.kt
@@ -1,4 +1,4 @@
-package com.nytimes.android.external.store3.multiplex
+package com.nytimes.android.external.store4.multiplex
 
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.ExperimentalCoroutinesApi
